requirements.txt
argparse
ipython
pyserial
